models/krate: Move `index_metadata()` fn into `index` module
